+++ packages/apollo-server-plugin-response-cache/src/ApolloServerPluginResponseCache.ts
@@ -86,6 +86,19 @@ interface Options<TContext = Record<string, any>> {
   shouldWriteToCache?(
     requestContext: GraphQLRequestContext<TContext>,
   ): ValueOrPromise<boolean>;
+
+  // This hook allows one to replace the function that is used to create a cache
+  // key. By default, it is the SHA-256 (from the Node `crypto` package) of the result of
+  // calling `JSON.stringify(keyData)`. You can override this to customize the serialization
+  // or the hash, or to make other changes like adding a prefix to keys to allow for
+  // app-specific prefix-based cache invalidation. You may assume that `keyData` is an object
+  // and that all relevant data will be found by the kind of iteration performed by
+  // `JSON.stringify`, but you should not assume anything about the particular fields on
+  // `keyData`.
+  generateCacheKey?(
+    requestContext: GraphQLRequestContext<Record<string, any>>,
+    keyData: unknown,
+  ): string;
 }
 
 enum SessionMode {
@@ -98,17 +111,27 @@ function sha(s: string) {
   return createHash('sha256').update(s).digest('hex');
 }
 
-interface BaseCacheKey {
+interface BaseCacheKeyData {
   source: string;
   operationName: string | null;
   variables: { [name: string]: any };
   extra: any;
 }
 
-interface ContextualCacheKey {
+interface ContextualCacheKeyData {
   sessionMode: SessionMode;
   sessionId?: string | null;
 }
+
+// We split the CacheKey type into two pieces just for convenience in the code
+// below. Note that we don't actually export this type publicly (the
+// generateCacheKey hook gets an `unknown` argument).
+type CacheKeyData = BaseCacheKeyData & ContextualCacheKeyData;
+
+type GenerateCacheKeyFunction = (
+  requestContext: GraphQLRequestContext<Record<string, any>>,
+  keyData: CacheKeyData,
+) => string;
 
 interface CacheValue {
   // Note: we only store data responses in the cache, not errors.
@@ -124,12 +147,6 @@ interface CacheValue {
   data: Record<string, any>;
   cachePolicy: Required<CacheHint>;
   cacheTime: number; // epoch millis, used to calculate Age header
-}
-
-type CacheKey = BaseCacheKey & ContextualCacheKey;
-
-function cacheKeyString(key: CacheKey) {
-  return sha(JSON.stringify(key));
 }
 
 function isGraphQLQuery(requestContext: GraphQLRequestContext<any>) {
@@ -148,8 +165,11 @@ export default function plugin(
         'fqc:',
       );
 
+      const generateCacheKey: GenerateCacheKeyFunction =
+        options.generateCacheKey ?? ((_, key) => sha(JSON.stringify(key)));
+
       let sessionId: string | null = null;
-      let baseCacheKey: BaseCacheKey | null = null;
+      let baseCacheKey: BaseCacheKeyData | null = null;
       let age: number | null = null;
 
       return {
@@ -163,12 +183,15 @@ export default function plugin(
           }
 
           async function cacheGet(
-            contextualCacheKeyFields: ContextualCacheKey,
+            contextualCacheKeyFields: ContextualCacheKeyData,
           ): Promise<GraphQLResponse | null> {
-            const key = cacheKeyString({
+            const cacheKeyData = {
               ...baseCacheKey!,
               ...contextualCacheKeyFields,
-            });
+            };
+
+            const key = generateCacheKey(requestContext, cacheKeyData);
+
             const serializedValue = await cache.get(key);
             if (serializedValue === undefined) {
               return null;
@@ -276,12 +299,15 @@ export default function plugin(
           }
 
           const cacheSetInBackground = (
-            contextualCacheKeyFields: ContextualCacheKey,
+            contextualCacheKeyFields: ContextualCacheKeyData,
           ): void => {
-            const key = cacheKeyString({
+            const cacheKeyData = {
               ...baseCacheKey!,
               ...contextualCacheKeyFields,
-            });
+            };
+
+            const key = generateCacheKey(requestContext, cacheKeyData);
+
             const value: CacheValue = {
               data,
               cachePolicy: policyIfCacheable,